Salt and Sacrifice: implement new game (#23)
* add salt_and_sanctuary implementation

Proposing an implementation of Salt and Sanctuary for Keymaster's Keep. This is a 2D Souls-like Metroidvania based around exploring new regions with equipment and completing bosses rushes by killing "Mages". This is intended to create a series of objectives to randomize kill order and occasionally the equipment required to do so.

* Update salt_and_sacrifice.py

fixed syntax issues